Populated the model with constants and currentweather obj
Now the project has a model to work with. I made a swift file that held the URL info constants.

I hard coded the latitude and longitude values for now because we have not pulled that info from the phone yet. I created the current weather object that encapsulates its private vars.