Grabs current location with permission and displays
This version finally grabs the current location of the user with his/her permission to give weather forecast. It fully displays the current weather for the day and the rest of the week based on the user's location. Now I have to fix up the MVC issues with the FORECAST being downloaded in the VC instead of its own model obj.